new structure and changes in components
