+++ components/Pagination/PaginationContent.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { cn } from '../../utils/utils';
+
+const PaginationContent = React.forwardRef<HTMLUListElement, React.ComponentProps<'ul'>>(
+  ({ className, ...props }, ref) => (
+    <ul ref={ref} className={cn('flex flex-row items-center gap-1', className)} {...props} />
+  )
+);
+PaginationContent.displayName = 'PaginationContent';
+
+export default PaginationContent;
