+++ components/Pagination/PaginationItem.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { cn } from '../../utils/utils';
+
+const PaginationItem = React.forwardRef<HTMLLIElement, React.ComponentProps<'li'>>(
+  ({ className, ...props }, ref) => <li ref={ref} className={cn('', className)} {...props} />
+);
+PaginationItem.displayName = 'PaginationItem';
+
+export default PaginationItem;
